+++ src/main/java/racingcar/MainApplication.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.Racing;
+
 public class MainApplication {
 	public static void main(String[] args) {
 		new Racing().playRace();
