+++ src/main/java/racingcar/controller/Racing.java
@@ -1,8 +1,17 @@
-package racingcar;
+package racingcar.controller;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
+
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
+import racingcar.domain.Condition;
+import racingcar.domain.NumberCondition;
+import racingcar.domain.RandomStrategy;
+import racingcar.util.InputView;
+import racingcar.util.Print;
+import racingcar.util.Views;
 
 public class Racing extends Condition {
 	private static final int MAX_OF_RANDOM = 10;
