+++ src/main/java/racingcar/domain/Condition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Condition {
 	private ConditionStrategy conditionStrategy;
