+++ src/main/java/racingcar/domain/ConditionStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface ConditionStrategy {
 	boolean validInput(String input) throws RuntimeException;
