+++ src/main/java/racingcar/util/Views.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.util;
+
+import racingcar.domain.Cars;
 
 public class Views {
