+++ src/test/java/racingcar/CarTest.java
@@ -7,6 +7,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import racingcar.domain.Car;
+import racingcar.domain.Position;
+
 public class CarTest {
 
 	@Test
