+++ src/test/java/racingcar/CarsTest.java
@@ -8,6 +8,10 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
+import racingcar.domain.Position;
+
 public class CarsTest {
 	@Test
 	@DisplayName("List<Car> 생성 - 일급 콜렉션")
