+++ src/test/java/racingcar/NameTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.Name;
+
 public class NameTest {
 	@Test
 	@DisplayName("이름 생성")
