+++ src/test/java/racingcar/RacingTest.java
@@ -7,6 +7,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import racingcar.controller.Racing;
+import racingcar.domain.Cars;
+
 public class RacingTest {
 	@Test
 	@DisplayName("레이싱 참가자 수 테스트")
